Create initial documentation
This captures the things that I did to set up the course.
I have created documents for the Course Content and Course Info
sections that corresdpond to the links in each section.
I don't know if we should switch the documentation to use markdown.
I have included the description text for Sections and Lectures
inside of parentheses.

I think it would be good for you to commit the Image when you have
created it.
I will commit the Promo Video as well when I have created it.