Add notes from the Udacity course
The notes follow more or less in order of the course. Sometimes I did place
some of the notes from later lectures in the same spot as I had started
taking notes in earlier lectures. At any rate, it's likely that we will
just cherry pick (no git pun intended) from the notes and use them as we
see fit.