Save initial high level terms
This script really is nowhere near a ready state. But, I wanted to
make sure to get it into the repo so that it's saved somewhere
remotely. I don't know why I'm feeling paranoid about messing something
up on my end and the need to get this stored remotely \O/